:recycle: refactor: loaders 분리
