+++ server/src/models/comment.ts
@@ -1,4 +1,4 @@
-import { pool } from "../db/index";
+import { pool } from "./../loaders/mysql";
 import { logger, BuildQuery } from "../utils";
 import { CommentIdDTO, CreateCommentDTO, GetCommentDTO, UpdateCommentDTO } from "../types";
 import { ResultSetHeader, RowDataPacket } from "mysql2";
