+++ server/src/sql/stickers.sql
