+++ server/src/sql/trigger.sql
